Modularize MTP benchmark orchestration and report validation
