+++ .github/workflows/ci.yml
@@ -17,6 +17,8 @@ jobs:
           python3 scripts/test-provider-release-resolution.py
           ./scripts/sync-install-embed.sh check
           ./scripts/test-prod-env-refresh.sh
+      - name: Test benchmark wrappers without models
+        run: make benchmark-wrapper-test
       - name: Test startup observer against local stubs
         run: python3 -m unittest discover -s scripts/startup_measurement -t scripts -p 'test_*.py'
