+++ scripts/mtp_benchmark/__init__.py
@@ -1,0 +1,1 @@
+"""Cache-only MTP benchmark orchestration and evidence validation."""
